third_party/WebKit/LayoutTests/media/controls/modern/scrubbing.html

<!DOCTYPE html>
<html>
<title>Test that player will behave correctly when scrubbing.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../media-controls.js"></script>
<video controls width=400 src="../../content/60_sec_video.webm"></video>
<script>
async_test(t => {
  const video = document.querySelector('video');
  const timeline = timelineElement(video);
  const timeElement = timelineThumbCurrentTime(video);
  const thumb = timelineThumb(video);

  let seeked = false;

  video.addEventListener('playing', t.step_func(() => {
    // Get the coordinates of the thumb and the timeline.
    const c = elementCoordinates(thumb);
    const c2 = elementCoordinates(timeline);

    // Add an event listener for when we start playing again after seeking.
    video.addEventListener('playing', t.step_func_done(() => {
      assert_true(seeked);
      checkControlsClassName(video, 'phase-ready state-playing');

      // Check the scrubbing UI is no longer shown.
      assert_equals('none', timeElement.style.display);
    }), { once: true });

    // Simulate a mouse down on those coordinates.
    chrome.gpuBenchmarking.pointerActionSequence([
      {
        source: 'mouse',
        actions: [
          { name: 'pointerMove', x: c[0], y: c[1] },
          { name: 'pointerDown', x: c[0], y: c[1], button: 'left' },
          { name: 'pointerMove', x: c2[0], y: c2[1]},
          { name: 'pause', duration: 5 },
          { name: 'pointerUp', button: 'left' }
        ]
      }
    ]);
  }), { once: true });

  video.addEventListener('seeking', t.step_func(() => {
    checkControlsClassName(video, 'phase-ready state-scrubbing');

    // Check the scrubbing UI is shown with the correct time.
    assert_equals('', timeElement.style.display);
    assert_equals('0:' + Math.round(video.currentTime),
        timeElement.textContent);
  }), { once: true });

  video.addEventListener('seeked', t.step_func(() => {
    seeked = true;
  }), { once: true });

  video.play();
});
</script>
</html>
